potential fix for larger datasets
error caused by integer overflow when calculating the number of values
in a volume. Switched everything to long unsigned int

This causes warnings in Volume.cpp though because there is a narrowing
conversion from long unsigned int to const int

+++ include/DataFile.h
@@ -26,10 +26,10 @@ namespace pbnj {
             std::string filename;
             FILETYPE filetype;
 
-            int xDim;
-            int yDim;
-            int zDim;
-            long int numValues;
+            unsigned long int xDim;
+            unsigned long int yDim;
+            unsigned long int zDim;
+            unsigned long int numValues;
 
             float minVal; // these should 
             float maxVal; //
